refactored structure. Added interface for discovery to implement multiple IaaS
