+++ pkg/output/output.go
@@ -1,7 +1,7 @@
 package output
 
-import "github.com/daspawnw/prometheus-aws-discovery/pkg/endpoints"
+import "github.com/daspawnw/prometheus-aws-discovery/pkg/discovery"
 
 type Output interface {
-	Write(endpoints.DiscoveredInstances) error
+	Write([]discovery.Instance) error
 }
