+++ pkg/output/output_file.go
@@ -1,9 +1,9 @@
-package outputfile
+package output
 
 import (
 	"io/ioutil"
 
-	"github.com/daspawnw/prometheus-aws-discovery/pkg/endpoints"
+	"github.com/daspawnw/prometheus-aws-discovery/pkg/discovery"
 	log "github.com/sirupsen/logrus"
 )
 
@@ -11,8 +11,8 @@ type OutputFile struct {
 	FilePath string
 }
 
-func (o OutputFile) Write(instances endpoints.DiscoveredInstances) error {
-	content, err := endpoints.ToJSONString(instances)
+func (o OutputFile) Write(instances []discovery.Instance) error {
+	content, err := discovery.TargetConfigBytes(instances)
 	if err != nil {
 		log.Error("Failed to convert instances to json string with error", err)
 		return err
